requirements.txt
aenum==3.1.16
appdirs==1.4.4
APScheduler==3.11.1
azure-core==1.36.0
azure-storage-blob==12.27.1
beautifulsoup4==4.14.2
bs4==0.0.2
cachetools==6.2.1
case-conversion==3.0.2
certifi==2025.10.5
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
clipboard==0.0.4
colorama==0.4.6
colour==0.1.5
cryptography==46.0.3
cursor==1.3.5
decorator==5.2.1
dill==0.4.0
fpdf==1.7.2
fuzzywuzzy==0.18.0
gender-guesser==0.4.0
google-api-core==2.28.1
google-api-python-client==2.186.0
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
googleapis-common-protos==1.71.0
html-text==0.7.1
html2text==2025.4.15
httplib2==0.31.0
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
IMAPClient==3.0.1
infi.systray==0.1.12.1
inflect==7.5.0
isodate==0.7.2
lxml==6.0.2
lxml_html_clean==0.4.3
maybe-else==0.2.1
mbstrdecoder==1.1.4
more-itertools==10.8.0
moviepy==2.2.1
msal==1.34.0
msoffcrypto-tool==5.4.2
numpy==2.3.4
o365==2.1.7
oauthlib==3.3.1
office365==0.3.15
Office365-REST-Python-Client==2.6.2
olefile==0.47
packaging==25.0
pandas==2.3.3
parsedatetime==2.6
pathmagic==0.3.14
pdf2image==1.17.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
prettierfier==1.0.3
proglog==0.1.12
proto-plus==1.26.1
protobuf==6.33.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydub==0.25.1
pyinstrument==5.1.1
pyiotools==0.3.18
PyJWT==2.10.1
pymiscutils==0.3.14
pyparsing==3.2.5
PyPDF2==3.0.1
pypdfium2==5.0.0
pyperclip==1.11.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.18
PyQt5_sip==12.17.1
pysubtypes==0.3.18
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
pyzmail36==1.0.5
readchar==4.2.1
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
schedule==1.2.2
Send2Trash==1.8.3
setuptools==80.9.0
simplejson==3.20.2
six==1.17.0
soupsieve==2.8
tabulate==0.9.0
tqdm==4.67.1
typeguard==4.4.4
typepy==1.3.4
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
xlsxwriter==3.2.9
